Rename file sort -> parse-results
